requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-WTF==1.1.1
Flask-Mail==0.9.1
python-dotenv==1.0.0
Pillow==10.0.0
bcrypt==4.0.1
PyJWT==2.8.0
markdown2==2.4.10
python-slugify==8.0.1
